pyproject.toml
[project]
name = "openhands-subagent-patterns"
version = "0.1.0"
description = "Stub POC demos for OpenHands subagent and orchestration patterns."
requires-python = ">=3.12"
dependencies = [
    "openhands-sdk>=1.12.0",
    "openhands-tools>=1.12.0",
    "openhands-workspace>=1.12.0",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
